+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - removed absolute font property declarations from CSS
 - added `A11yMenu` supporting `medium` and larger displays
+- use `FontsLoader` component
 
 ## `@svizzle/ui` v0.5.0 (next)
 
@@ -21,6 +22,7 @@
 	- removed `fontSize` property
 - add `StorageIO` component
 - add `A11yMenu` component
+- add `FontsLoader` component
 
 ## `@svizzle/atlas` v0.6.0 (next)